Update NewLogoVector.svg
this is actually working. I can see now that i might want to add the "Resume" text to the SVG file. Until i am able to do that, i will do it like this.

Maybe it is better to add the logo to the jumbotron. in a seperate column or something. It is something to think about